Refactor configure step for compilation
Rather than relying on users to specify that they have SASL, we can use
librdkafka's built in checking to find out. If we parse out the header
file that is made we can export environment variables so node-gyp can
configure itself appropriately.

This commit should also improve development because configure is now run
at install time only. It means when developing you do not need to
recompile the entirety of librdkafka with each minor change